+++ src/main/java/com/charbel/ecommerce/orders/dto/OrderResponse.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -21,10 +22,10 @@ public class OrderResponse {
 	private String userEmail;
 	private String userFirstName;
 	private String userLastName;
-	private Integer originalAmount;
-	private Integer discountAmount;
-	private Integer deliveryFee;
-	private Integer totalAmount;
+	private BigDecimal originalAmount;
+	private BigDecimal discountAmount;
+	private BigDecimal deliveryFee;
+	private BigDecimal totalAmount;
 	private OrderStatus status;
 	private List<OrderItemResponse> orderItems;
 	private LocalDateTime createdAt;
